Custom Bullet Movement
Default UE4 bullet physics removed and replaced with custom system. Bullet moves based on "speed" in the direction of fire and interacts "penetration" number of objects.

Bug:
Bullet only seems to interact with objects that have "Simulate Physics" enabled. Although this is not desirable and should be investigated, this issue can be ignored for now if we assume all objects will be simulating physics.