Updated playing bounds, limitted fire rate, promoted ship mesh to root so collision occurs.
Bounds now uses and inverted cube model and simplified settings for size.
Guns have a new float for cool down between shots, and toggles a bool when ready to shoot again.
Ship mesh had to be made root to allow physics interactions but did avoid needing the ship to be physics driven.